==> notebook/01-preprocessing_and_parse_faq.ipynb <==
### Download the dataset

[comprehensive-medical-q-a-dataset](https://www.kaggle.com/datasets/thedevastator/comprehensive-medical-q-a-dataset)\
The MedQuad dataset provides a comprehensive source of medical questions and answers for natural language processing. With over 43,000 patient inquiries from real-life situations categorized into 31 distinct types of questions, the dataset offers an invaluable opportunity to research correlations between treatments, chronic diseases, medical protocols and more. Answers provided in this database come not only from doctors but also other healthcare professionals such as nurses and pharmacists, providing a more complete array of responses to help researchers unlock deeper insights within the realm of healthcare. 

# download the dataset from kaggle

import kagglehub

# Download latest version
path = kagglehub.dataset_download("thedevastator/comprehensive-medical-q-a-dataset")

print("Path to dataset files:", path)

import json
import re
import pandas as pd
from tqdm.auto import tqdm

# Load dataset
df_raw = pd.read_csv("./../dataset/medical_qa_raw.csv")
df_raw.head(5)

df_raw.describe()

# Shape and columns
print("Shape:", df_raw.shape)
print("Columns:", df_raw.columns.tolist())

# Null values check
print("\nMissing Values:")
print(df_raw.isnull().sum())

# Function to clean spaces, tabs, and normalize text
def clean_text(text):
    if pd.isna(text):
        return ""
    text = str(text).replace("\t", " ")             # replace tabs with spaces
    text = re.sub(r"\s+", " ", text)                # collapse multiple spaces
    text = re.sub(r"\s+\?", "?", text)              # remove space before ?
    text = re.sub(r"\?+", "?", text)                # replace ?? or ??? with single ?
    return text.strip()

df_clean = df_raw.copy()
# Apply cleaning to all string columns
for col in df_clean.select_dtypes(include=["object"]).columns:
    df_clean[col] = df_clean[col].apply(clean_text)
  

def remove_duplicate_questions(df):
    """Remove exact duplicate questions"""
    print(f"Original dataset shape: {df.shape}")
    
    # Remove exact duplicates based on Question column
    df_clean = df.drop_duplicates(subset=['Question'], keep='first')
    
    print(f"After removing duplicate questions: {df_clean.shape}")
    print(f"Removed {len(df) - len(df_clean)} duplicate questions")
    
    return df_clean

# Basic duplicate removal
df_unique = remove_duplicate_questions(df_clean)

df_unique.describe()

#### generate unique id for each record

import hashlib
def generate_document_id(doc):
    # combined = f"{doc['course']}-{doc['question']}"
    combined = f"{doc['Question']}-{doc['Answer'][:15]}"
    hash_object = hashlib.md5(combined.encode())
    hash_hex = hash_object.hexdigest()
    document_id = hash_hex[:8]
    return document_id

df_unique["id"] = df_unique.apply(lambda row: generate_document_id(row), axis=1)

# Save cleaned CSV
df_unique.to_csv("./../dataset/medical_qa_with_id.csv", index=False)

df_unique

### Save the documents as a json file

documents = []

for  _, row in df_unique.iterrows():
    doc = {
        "answer": row["Answer"].strip(),
        "question": row["Question"].strip(),
        "qtype": row['qtype'].strip(),
        "id": row['id']

    }
    documents.append(doc)

final_data = [
    {
        "document_info": "Comprehensive Medical Q&A Dataset",
        "documents": documents
    }
]

# Save JSON file
with open("./../dataset/medical_qa_documents_with_id.json", "w", encoding="utf-8") as f_in:
    json.dump(final_data, f_in, indent=2, ensure_ascii=False)

print("Saved cleaned JSON at ./../dataset/medical_qa_documents_with_id.json")


!head ./../dataset/medical_qa_documents_with_id.json
