==> notebook/04a-generate_embedding_vector.ipynb <==
import json
import pickle
import numpy as np
from tqdm.auto import tqdm
from sentence_transformers import SentenceTransformer

with open('./../dataset/medical_qa_documents_with_id.json', 'rt') as f_in:
    docs_with_id = json.load(f_in)

documents = []

for docs_info in docs_with_id:
    for doc in docs_info['documents']:
        documents.append(doc)

### loading embedding model

# embdedding model
model_name = 'multi-qa-MiniLM-L6-cos-v1'
model = SentenceTransformer(model_name)

### Generate embedding vectors

vectors = []

for doc in tqdm(documents):
    question = doc['question']
    answer = doc['answer']
    vector = model.encode(question + ' ' + answer)
    vectors.append(vector)

#### saved embedding as a pickle

# Save
with open("./../dataset/documents-vectors-multi-qa-MiniLM-L6-cos-v1.pkl", "wb") as f:
    pickle.dump(vectors, f)

print("Saved vectors locally")

# # Load later
# with open("cos-v1-model-vectors.pkl", "rb") as f:
#     vectors = pickle.load(f)

#### Saved embedding as a Json File

# Convert numpy arrays to lists
vectors_json = [v.tolist() for v in vectors]

with open("./../dataset/documents-vectors-multi-qa-MiniLM-L6-cos-v1.json", "w") as f:
    json.dump(vectors_json, f)

print("Saved as JSON")


# load later
# with open("vectors.json", "r") as f:
#     vectors = json.load(f)
# vectors = [np.array(v) for v in vectors]
