==> notebook/04b-indexing_with_vector_search_qdrant_and_rag_flow.ipynb <==
import json
from tqdm import tqdm
import pickle

### Data Ingestion in Qdrant


# Load dataset from JSON
with open("./../dataset/medical_qa_documents_with_id.json", "r") as f:
    docs_raw = json.load(f)

documents = docs_raw[0]["documents"]  # get the list of Q&A documents

print("Total documents loaded:", len(documents))
print("Keys in document:", docs_raw[0].keys())


documents = []

for docs_info in docs_raw:
    for doc in docs_info['documents']:
        documents.append(doc)

documents[0]

from qdrant_client import QdrantClient, models

qd_client = QdrantClient(
    url="http://localhost:6333",
    timeout=60  # seconds (1 minutes)
)

collection_name = "medical-faq"

# delete collection name if already created in qdrant
qd_client.delete_collection(collection_name=collection_name)

#### generating embedding vector

from sentence_transformers import SentenceTransformer

EMBEDDING_DIMENSIONALITY = 384  # for multi-qa-MiniLM-L6-cos-v1

model_name = 'multi-qa-MiniLM-L6-cos-v1'
model = SentenceTransformer(model_name)

### Generate embedding vectors

# no need to run embedding is already generated and locally saved
vectors = []

for doc in tqdm(documents):
    question = doc['question']
    answer = doc['answer']
    vector = model.encode(question + ' ' + answer)
    vectors.append(vector)

### Load the embedding vector if already generated

with open("./../dataset/documents-vectors-multi-qa-MiniLM-L6-cos-v1.pkl", "rb") as f:
    vectors = pickle.load(f)

print(f"Loaded {len(vectors)} vectors and {len(documents)} documents")

qd_client.recreate_collection(
    collection_name=collection_name,
    vectors_config=models.VectorParams(
        size=EMBEDDING_DIMENSIONALITY,
        distance=models.Distance.COSINE
    )
)

qd_client.create_payload_index(
    collection_name=collection_name,
    field_name="qtype",
    field_schema="keyword"
)

from tqdm.auto import tqdm
import psutil, os, gc, time

process = psutil.Process(os.getpid())
BATCH_SIZE = 100

for start in tqdm(range(0, len(vectors), BATCH_SIZE), desc="Uploading to Qdrant"):
    end = start + BATCH_SIZE
    batch_vectors = vectors[start:end]
    batch_docs = documents[start:end]

    points = []
    for i, (vec, doc) in enumerate(zip(batch_vectors, batch_docs)):
        point = models.PointStruct(
            id=start + i,
            vector=vec.tolist(),
            payload=doc
        )
        points.append(point)

    qd_client.upsert(collection_name=collection_name, points=points)
    gc.collect()

    mem = process.memory_info().rss / 1e6
    print(f"Uploaded {end}/{len(vectors)} | Memory: {mem:.2f} MB")
    time.sleep(0.1)


## RAG with Vector Search


import os
os.environ["GROQ_API_KEY"] = "gsk_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"

from groq import Groq

client = Groq()

def build_prompt(query, search_results):

    prompt_template = """
You are a professional medical assistant.
Answer the QUESTION using only the CONTEXT provided from verified medical sources.
If the answer is not available in the CONTEXT, say "I'm not sure based on the available information."

QUESTION: {question}

CONTEXT:
{context}
""".strip()
    
    context = ""
    
    for doc in search_results:
        context = context + f"question: {doc['question']}\nanswer: {doc['answer']}\n\n"
    
    prompt = prompt_template.format(question=query, context=context).strip()
    return prompt

def llm(prompt):
    response = client.chat.completions.create(
        model="openai/gpt-oss-20b",
        messages=[{"role": "user", "content": prompt}]
    )
    
    return response.choices[0].message.content

def vector_search(query, qtype=None, limit=5):
    query_vector = model.encode([query])[0].tolist()

    query_filter = None
    if qtype:
        query_filter = models.Filter(
            must=[models.FieldCondition(key="qtype", match=models.MatchValue(value=qtype))]
        )

    results = qd_client.query_points(
        collection_name=collection_name,
        query=query_vector,
        query_filter=query_filter,
        limit=limit,
        with_payload=True
    )

    return [p.payload for p in results.points]


def rag(query, qtype):
    search_results = vector_search(query, qtype)
    prompt = build_prompt(query, search_results)
    answer = llm(prompt)
    return answer

query = "what is malaria?"

qtype = 'information'
rag(query, qtype)
