==> notebook/05b-retrieval_evaluation_elastic_qdrant_and_minsearch.ipynb <==
import json
from tqdm.auto import tqdm
import pandas as pd

with open('./../dataset/medical_qa_documents_with_id.json', 'rt') as f_in:
    docs_raw = json.load(f_in)

documents = []
for docs_info in docs_raw:
    for doc in docs_info['documents']:
        documents.append(doc)

df_ground_truth = pd.read_csv('./../dataset/search_ground-truth-data.csv')

ground_truth_dict = df_ground_truth.to_dict(orient='records')

### Elastic Search Evaluation

from elasticsearch import Elasticsearch

es_client = Elasticsearch(
    ["http://localhost:9200"],   # must include scheme http://
    request_timeout=60           # increase timeout in case ES is slow
)
print(es_client.info)

index_settings = {
    "settings": {
        "number_of_shards": 1,
        "number_of_replicas": 0
    },
    "mappings": {
        "properties": {
            "answer": {"type": "text"},
            "question": {"type": "text"},
            "qtype": {"type": "keyword"},
            "id": {"type": "keyword"}

        }
    }
}

index_name = "medical-questions"

es_client.indices.delete(index=index_name, ignore_unavailable=True)
es_client.indices.create(index=index_name, body=index_settings)

for doc in tqdm(documents):
    es_client.index(index=index_name, document=doc)

def elastic_search(query, qtype):
    search_query = {
        "size": 5,
        "query": {
            "bool": {
                "must": {
                    "multi_match": {
                        "query": query,
                        "fields": ["question^3", "answer", "qtype"],
                        "type": "best_fields"
                    }
                },
                "filter": {
                    "term": {
                        "qtype": qtype
                    }
                }
            }
        }
    }

    response = es_client.search(index=index_name, body=search_query)
    
    result_docs = []
    
    for hit in response['hits']['hits']:
        result_docs.append(hit['_source'])
    
    return result_docs

elastic_search(
    query = "what is malaria?",
    qtype="information"
)

relevance_total = []

for q in tqdm(ground_truth_dict):
    doc_id = q['document']
    results = elastic_search(query=q['question'], qtype=q['qtype'])
    relevance = [d['id'] == doc_id for d in results]
    relevance_total.append(relevance)

def hit_rate(relevance_total):
    cnt = 0

    for line in relevance_total:
        if True in line:
            cnt = cnt + 1

    return cnt / len(relevance_total)

def mrr(relevance_total):
    total_score = 0.0

    for line in relevance_total:
        for rank in range(len(line)):
            if line[rank] == True:
                total_score = total_score + 1 / (rank + 1)

    return total_score / len(relevance_total)

- hit-rate (recall)
- Mean Reciprocal Rank (mrr)

hit_rate(relevance_total), mrr(relevance_total)

### MinSearch Evaluation

import minsearch

index = minsearch.Index(
    text_fields=["question", "answer"],
    keyword_fields=["qtype", "id"]
)

index.fit(documents)

def minsearch_search(query, qtype):
    boost = {'question': 3.0}

    results = index.search(
        query=query,
        filter_dict={'qtype': qtype},
        boost_dict=boost,
        num_results=5
    )

    return results

relevance_total = []

for q in tqdm(ground_truth_dict):
    doc_id = q['document']
    results = minsearch_search(query=q['question'], qtype=q['qtype'])
    relevance = [d['id'] == doc_id for d in results]
    relevance_total.append(relevance)

hit_rate(relevance_total), mrr(relevance_total)

### Qdrant Vector Search Evaluation

from qdrant_client import QdrantClient, models

qd_client = QdrantClient(
    url="http://localhost:6333",
    timeout=60  # seconds (1 minutes)
)

collection_name = "medical-faq"

from sentence_transformers import SentenceTransformer

EMBEDDING_MODEL_NAME = "multi-qa-MiniLM-L6-cos-v1"
model = SentenceTransformer(EMBEDDING_MODEL_NAME)

def vector_search(qd_client, query, qtype=None, limit=5):
    query_vector = model.encode([query])[0].tolist()

    query_filter = None
    if qtype:
        query_filter = models.Filter(
            must=[models.FieldCondition(key="qtype", match=models.MatchValue(value=qtype))]
        )

    results = qd_client.query_points(
        collection_name=collection_name,
        query=query_vector,
        query_filter=query_filter,
        limit=limit,
        with_payload=True
    )

    return [p.payload for p in results.points]

relevance_total = []

for q in tqdm(ground_truth_dict):
    doc_id = q['document']
    results = vector_search(qd_client, query=q['question'], qtype=q['qtype'])
    relevance = [d['id'] == doc_id for d in results]
    relevance_total.append(relevance)

hit_rate(relevance_total), mrr(relevance_total)

def evaluate(ground_truth, search_function):
    relevance_total = []

    for q in tqdm(ground_truth):
        doc_id = q['document']
        results = search_function(q)
        relevance = [d['id'] == doc_id for d in results]
        relevance_total.append(relevance)

    return {
        'hit_rate': hit_rate(relevance_total),
        'mrr': mrr(relevance_total),
    }

# Elastic Search Evaluation
evaluate(ground_truth_dict, lambda q: elastic_search(q['question'], q['qtype']))

# MinSearch Evaluation
evaluate(ground_truth_dict, lambda q: minsearch_search(q['question'], q['qtype']))

# Qdrant VectorSearch Evaluation
evaluate(ground_truth_dict, lambda q: vector_search(qd_client,q['question'], q['qtype']))
